Drop PHP 7.1 support.
