+++ src/test/php/PDepend/Source/AST/ASTLiteralTest.php
@@ -172,7 +172,7 @@ class ASTLiteralTest extends ASTNodeTest
      *
      * @return void
      * @since 1.0.0
-     * @covers \PDepend\Source\Language\PHP\PHPParserVersion71
+     * @covers \PDepend\Source\Language\PHP\PHPParserVersion72
      */
     public function testLiteralWithZeroBinaryIntegerValue()
     {
@@ -209,7 +209,7 @@ class ASTLiteralTest extends ASTNodeTest
      *
      * @return void
      * @since 1.0.0
-     * @covers \PDepend\Source\Language\PHP\PHPParserVersion71
+     * @covers \PDepend\Source\Language\PHP\PHPParserVersion72
      */
     public function testLiteralWithNonZeroBinaryIntegerValue()
     {
